+++ src/acceptance/org/smoothbuild/acceptance/slib/java/JunitTest.java
@@ -14,8 +14,8 @@ public class JunitTest extends AcceptanceTestCase {
     createJunitLibs();
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createUserModule("""
-            junitJars = files("junit");
-            srcJar = files("src") | javac(libs=junitJars) | jar;                                  
+            junitJars = projectFiles("junit");
+            srcJar = projectFiles("src") | javac(libs=junitJars) | jar;                                  
             result = junit(tests=srcJar, deps=[]);                  
             """);
     runSmoothBuild("result");
@@ -28,8 +28,8 @@ public class JunitTest extends AcceptanceTestCase {
     createJunitLibs();
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createUserModule("""
-            junitJars = files("junit");
-            srcJar = files("src") | javac(libs=junitJars) | jar;
+            junitJars = projectFiles("junit");
+            srcJar = projectFiles("src") | javac(libs=junitJars) | jar;
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -41,8 +41,8 @@ public class JunitTest extends AcceptanceTestCase {
     createJunitLibs();
     createFile("src/" + FAILING_TEST_CLASS + ".java", failingTestSourceCode());
     createUserModule("""
-            junitJars = files("junit");
-            srcJar = files("src") | javac(libs=junitJars) | jar;
+            junitJars = projectFiles("junit");
+            srcJar = projectFiles("src") | javac(libs=junitJars) | jar;
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -55,8 +55,8 @@ public class JunitTest extends AcceptanceTestCase {
     createJunitLibs();
     createDir("src");
     createUserModule("""
-            junitJars = files("junit");
-            srcJar = files("src") | javac(libs=junitJars) | jar;
+            junitJars = projectFiles("junit");
+            srcJar = projectFiles("src") | javac(libs=junitJars) | jar;
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -70,8 +70,8 @@ public class JunitTest extends AcceptanceTestCase {
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createFile("src/" + FAILING_TEST_CLASS + ".java", failingTestSourceCode());
     createUserModule(format("""
-            junitJars = files("junit");
-            srcJar = files("src") | javac(libs=junitJars) | jar;
+            junitJars = projectFiles("junit");
+            srcJar = projectFiles("src") | javac(libs=junitJars) | jar;
             result = junit(include="%s", tests=srcJar, deps=junitJars);
             """, SUCCESSFUL_TEST_CLASS));
     runSmoothBuild("result");